requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests

msrest == 0.6.18
msrestazure == 0.6.4
azure-core == 1.7.0
azure-common == 1.1.26
azure-identity == 1.4.0
azure-mgmt-core == 1.2.0
azure-storage-blob == 12.4.0
azure-keyvault-secrets == 4.2.0  